requirements.txt
aiogram==3.10.0
aiohttp==3.9.1
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0